+++ src/components/Modal/ModalSignIn.module.css
@@ -25,7 +25,7 @@
   padding-right: 5px;
 }
 
-button {
+/* button {
   display: flex;
   background-color: var(--green);
   align-items: center;
@@ -33,7 +33,7 @@ button {
   height: 40px;
   align-self: center;
   cursor: pointer;
-}
+} */
 p {
   margin: 0;
   padding: 0;
