+++ src/components/Navigation/Drawer/Drawer.module.css
@@ -23,11 +23,17 @@
 }
 
 .Drawer ul {
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   list-style: none;
   margin: 0;
   padding: 0;
 }
 .Drawer ul li {
+  width: 80%;
+  float: left;
   margin-bottom: 15px;
 }
 .Drawer ul li a {
