+++ src/components/UI/Buttons/Button/Button.module.css
@@ -15,7 +15,7 @@
   text-transform: uppercase;
   font-weight: bold;
   font-size: 12px;
-}
+}  
 
 .Button:focus {
   outline: none;
