+++ src/pages/TableStyle.module.css
@@ -1,124 +1,135 @@
 .TableStyle {
-    display: flex;
-    justify-content: center;
+  display: flex;
+  justify-content: center;
 }
 
 .TableStyle ul {
-    height: auto;
+  height: auto;
 }
 
 .TableStyle ul {
-    height: 100%; 
+  height: 100%;
 }
 
 .TableStyle ul li {
-    padding: 10px;
-    display: flex;
-    width: 100%;
-    height: 100%;
-    border: 1px solid black;
-    align-items: flex-start;
-    justify-content: center;
+  padding: 10px;
+  display: flex;
+  width: 100%;
+  height: 100%;
+  border: 1px solid black;
+  align-items: flex-start;
+  justify-content: center;
 }
 
 .TableStyle ul li:first-child {
-    width: 100px;
+  width: 100px;
 }
 
 .TableStyle ul li {
-    height: auto;
-    padding: 0;
+  height: auto;
+  padding: 0;
 }
 
 ul {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    justify-content: space-between;
-    list-style: none;
-    padding: 0;
-    margin: 0;
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  justify-content: space-between;
+  list-style: none;
+  padding: 0;
+  margin: 0;
 }
 
 .header {
-    background-color: green;
+  background-color: green;
 }
 
-
 p {
-    display: inline-block;
-    word-break: break-all;
+  display: inline-block;
+  word-break: break-all;
 }
 
 .table {
-    height: 35px;
-    background-color: #fff;
+  height: 35px;
+  background-color: #fff;
 }
 
-.table li {
-    height: auto;
+.TableStyle .table li {
+  display: flex;
+  align-items: center;
+  min-height: 30px;
+  height: auto;
 }
 
 button {
-    width: 4.5rem;
-    height: 40px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border-radius: 4px;
-    outline: none;
-    background-color: green;
-    border: none;
-    color: #fff;
-    margin-top: 5px;
-    margin-bottom: 5px;
-    margin-left: 10px;
+  width: 4.5rem;
+  height: 40px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  border-radius: 4px;
+  outline: none;
+  background-color: green;
+  border: none;
+  color: #fff;
+  margin-top: 5px;
+  margin-bottom: 5px;
+  margin-left: 0px;
+  border: 1px solid black;
+  cursor: pointer;
 }
 
 .button .fontButton {
-    font-size: 14px;
-    
+  font-size: 14px;
 }
-.registration{
-    align-self: flex-start;
-    margin-left: 0px;
+.registration {
+  align-self: flex-start;
+  margin-left: 0px;
 }
 
 .delete {
-    background-color: red;
+  background-color: red;
 }
 
 a {
-    text-decoration: none;
+  text-decoration: none;
 }
 
 .TableStyle ul .actions {
-    justify-content: space-around;
-    align-items: flex-start;
-    display: flex;
-    flex-wrap: wrap;
-    width: 100%;
+  justify-content: space-around;
+  align-items: flex-start;
+  display: flex;
+  flex-wrap: wrap;
+  width: 100%;
 }
 
 .TableStyle ul .age {
-    width: 150px;
+  width: 150px;
 }
 
 .TableStyle ul .education {
-    width: 400px;
+  width: 400px;
 }
 
-.warning{
-    background-color: #F0AD4E;
+.warning {
+  background-color: #f0ad4e;
 }
-.danger{
-    background-color: #D9534F;
+.danger {
+  background-color: #d9534f;
 }
-.default{
-    background-color: #0275D8;;
+.default {
+  background-color: #0275d8;
 }
-i{
-    font-style: normal;
-    outline: none;
-    text-decoration: underline;
+.back {
+  background-color: #ccc;
+  align-self: flex-end;
+  color: black;
+  width: auto;
+  min-width: 4.5rem;
+
+}
+i {
+  font-style: normal;
+  outline: none;
+  text-decoration: underline;
 }
